+++ java/component/keyple-plugin/stub/src/main/java/org/eclipse/keyple/plugin/stub/StubPlugin.java
@@ -47,7 +47,7 @@ public final class StubPlugin extends AbstractThreadedObservablePlugin {
          * Monitoring is not handled by a lower layer (as in PC/SC), reduce the threading period to
          * 10 ms to speed up responsiveness.
          */
-        threadWaitTimeout = 1000;
+        threadWaitTimeout = 10;
     }
 
     /**
