+++ java/component/keyple-plugin/stub/src/test/java/org/eclipse/keyple/plugin/stub/StubPluginTest.java
@@ -62,29 +62,7 @@ public class StubPluginTest extends BaseStubTest {
     }
 
 
-    /**
-     * Plug one reader asynchronously Check: Event thrown
-     */
-    @Test
-    public void plugOneReaderAsync_success() throws InterruptedException, KeypleReaderException {
-        final CountDownLatch lock = new CountDownLatch(1);
-        final String READER_NAME = "plugOneReaderAsync_sucess";
 
-        // add READER_CONNECTED assert observer
-        stubPlugin.addObserver(new ObservablePlugin.PluginObserver() {
-            @Override
-            public void update(PluginEvent event) {
-                Assert.assertEquals(PluginEvent.EventType.READER_CONNECTED, event.getEventType());
-                Assert.assertEquals(1, event.getReaderNames().size());
-                Assert.assertEquals(READER_NAME, event.getReaderNames().first());
-                lock.countDown();
-            }
-        });
-
-        stubPlugin.plugStubReader(READER_NAME, TransmissionMode.CONTACTLESS, false);
-        lock.await(2, TimeUnit.SECONDS);
-        Assert.assertEquals(0, lock.getCount());
-    }
 
     /**
      * Plug one reader synchronously Check: Event thrown
@@ -129,55 +107,7 @@ public class StubPluginTest extends BaseStubTest {
 
     }
 
-    /**
-     * Unplug one reader and wait for event
-     */
-    @Test
-    public void unplugOneReaderAsync_success() throws InterruptedException, KeypleReaderException {
-        final CountDownLatch connectedLock = new CountDownLatch(1);
-        final CountDownLatch disconnectedLock = new CountDownLatch(1);
-        final String READER_NAME = "unplugOneReaderAsync_success";
 
-        ObservablePlugin.PluginObserver disconnected_obs = new ObservablePlugin.PluginObserver() {
-            int event_i = 1;
-
-            @Override
-            public void update(PluginEvent event) {
-                logger.info("event {} #readers {}", event.getEventType(),
-                        event.getReaderNames().size());
-                // first event, should be a READER_CONNECTED
-                if (event_i == 1) {
-                    Assert.assertEquals(PluginEvent.EventType.READER_CONNECTED,
-                            event.getEventType());
-                    connectedLock.countDown();
-                }
-                // analyze the second event, should be a READER_DISCONNECTED
-                if (event_i == 2) {
-                    Assert.assertEquals(PluginEvent.EventType.READER_DISCONNECTED,
-                            event.getEventType());
-                    Assert.assertEquals(1, event.getReaderNames().size());
-                    Assert.assertEquals(READER_NAME, event.getReaderNames().first());
-                    disconnectedLock.countDown();
-                }
-                event_i++;
-            }
-        };
-
-        // add ReaderEvent observer
-        stubPlugin.addObserver(disconnected_obs);
-
-        // plug a reader
-        stubPlugin.plugStubReader(READER_NAME, false);
-
-        connectedLock.await(2, TimeUnit.SECONDS);
-
-        // unplug reader
-        stubPlugin.unplugStubReader(READER_NAME, false);
-
-        // wait for event to be raised
-        disconnectedLock.await(2, TimeUnit.SECONDS);
-        Assert.assertEquals(0, disconnectedLock.getCount());
-    }
 
     /**
      * Plug same reader twice Check : only one reader
@@ -202,7 +132,9 @@ public class StubPluginTest extends BaseStubTest {
     }
 
     /**
-     * Plug many readers at once sync Check : count readers
+     * Plug many readers at once sync
+     *
+     * Check : count readers
      */
     @Test
     public void plugMultiReadersSync_success() throws InterruptedException, KeypleReaderException {
@@ -215,40 +147,7 @@ public class StubPluginTest extends BaseStubTest {
         Assert.assertEquals(3, stubPlugin.getReaders().size());
     }
 
-    /**
-     * Plug many readers at once async Check : check event, count in event
-     */
-    @Test
-    public void plugMultiReadersASync_success() throws InterruptedException, KeypleReaderException {
-        final Set<String> READERS =
-                new HashSet<String>(Arrays.asList("E_Reader1", "E_Reader2", "E_Reader3"));
 
-        // lock test until message is received
-        final CountDownLatch readerConnected = new CountDownLatch(1);
-
-        // add READER_CONNECTED assert observer
-        stubPlugin.addObserver(new ObservablePlugin.PluginObserver() {
-            @Override
-            public void update(PluginEvent event) {
-                logger.info("event {} #readers {}", event.getEventType(),
-                        event.getReaderNames().size());
-                Assert.assertEquals(PluginEvent.EventType.READER_CONNECTED, event.getEventType());
-                Assert.assertEquals(3, event.getReaderNames().size());
-                Assert.assertEquals(READERS, event.getReaderNames());
-                readerConnected.countDown();
-            }
-
-        });
-
-
-        // connect readers
-        stubPlugin.plugStubReaders(READERS, false);
-
-        // wait for event to be raised
-        readerConnected.await(2, TimeUnit.SECONDS);
-        Assert.assertEquals(0, readerConnected.getCount());
-
-    }
 
     /**
      * Plug and unplug many readers at once synchronously Check : count
@@ -265,62 +164,5 @@ public class StubPluginTest extends BaseStubTest {
         Assert.assertEquals(0, stubPlugin.getReaders().size());
     }
 
-    /**
-     * Plug and unplug many readers at once asynchronously Check : event and count events
-     */
-    @Test
-    public void plugUnplugMultiReadersAsync_success()
-            throws InterruptedException, KeypleReaderException {
-        final Set<String> READERS =
-                new HashSet<String>(Arrays.asList("F_Reader1", "F_Reader2", "F_Reader3"));
 
-        // lock test until message is received
-        final CountDownLatch connectedLock = new CountDownLatch(1);
-        final CountDownLatch disconnectedLock = new CountDownLatch(1);
-
-        ObservablePlugin.PluginObserver assertDisconnect = new ObservablePlugin.PluginObserver() {
-            int event_i = 1;
-
-            @Override
-            public void update(PluginEvent event) {
-                logger.info("event {} #readers {}", event.getEventType(),
-                        event.getReaderNames().size());
-                if (event_i == 1) {
-                    Assert.assertEquals(PluginEvent.EventType.READER_CONNECTED,
-                            event.getEventType());
-                    connectedLock.countDown();
-                }
-                // analyze the second event, should be a READER_DISCONNECTED
-                if (event_i == 2) {
-                    Assert.assertEquals(PluginEvent.EventType.READER_DISCONNECTED,
-                            event.getEventType());
-                    // Assert.assertEquals(3, event.getReaderNames().size());
-                    Assert.assertEquals(READERS, event.getReaderNames());
-                    disconnectedLock.countDown();
-                }
-                event_i++;
-            }
-        };
-        // add assert DISCONNECT assert observer
-        stubPlugin.addObserver(assertDisconnect);
-
-        // connect reader
-        stubPlugin.plugStubReaders(READERS, false);
-
-        Assert.assertTrue(connectedLock.await(2, TimeUnit.SECONDS));
-
-        Thread.sleep(1000);
-
-        stubPlugin.unplugStubReaders(READERS, false);
-
-        Assert.assertTrue(disconnectedLock.await(2, TimeUnit.SECONDS));
-
-        // Thread.sleep(1000);
-
-        Thread.sleep(1000);
-        logger.debug("Stub Readers connected {}", stubPlugin.getReaderNames());
-        Assert.assertEquals(0, stubPlugin.getReaders().size());
-        Assert.assertEquals(0, connectedLock.getCount());
-        Assert.assertEquals(0, disconnectedLock.getCount());
-    }
 }
